# api/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "description": "This is api design for binar full-stack assessment test",
      "version": "1.0.0",
      "title": "Swagger Movie DB API",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
         "email": "apiteam@swagger.io"
      },
      "license": {
         "name": "Apache 2.0",
         "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
...
   "paths": {
...
      "/users/{userId}": {
         "get": {
            "tags": [
               "users"
            ],
            "summary": "Get detail user data in particular userId",
            "operationId": "getUserByID",
            "parameters": [
               {
                  "name": "userId",
                  "in": "path",
                  "description": "ID of user to retrieve specific user data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "in": "header",
                  "name": "Authorization",
                  "description": "Authorization bearer token",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid user ID"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "User not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "put": {
            "tags": [
               "users"
            ],
            "summary": "Update specific user data according to user ID",
            "operationId": "updateUserByID",
            "parameters": [
               {
                  "name": "userId",
                  "in": "path",
                  "description": "ID of user to retrieve specific user data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Authorization",
                  "in": "header",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "requestBody": {
               "description": "Updated user object",
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "anyOf": [
                           {
                              "$ref": "#/components/schemas/User"
                           }
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid user ID"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "User not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "delete": {
            "tags": [
               "users"
            ],
            "summary": "Delete user data based on defined ID",
            "description": "",
            "operationId": "deleteUser",
            "parameters": [
               {
                  "in": "header",
                  "name": "Authorization",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               },
               {
                  "name": "userId",
                  "in": "path",
                  "description": "ID of user to retrieve specific user data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid supplied ID"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "User not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "User": {
            "type": "object",
            "properties": {
               "user_id": {
                  "type": "string",
                  "format": "uuid"
               },
               "num": {
                  "type": "integer",
                  "format": "int32"
               },
               "email": {
                  "type": "string",
                  "format": "email"
               },
               "password": {
                  "type": "string"
               },
               "level_id": {
                  "type": "integer",
                  "format": "int64",
                  "description": "User authorization level (external/admin)",
                  "anyOf": [
                     {
                        "$ref": "#/components/schemas/Level"
                     }
                  ]
               },
               "created_at": {
                  "type": "string",
                  "format": "date-time"
               },
               "updated_at": {
                  "type": "string",
                  "format": "date-time"
               },
               "deleted_at": {
                  "type": "string",
                  "format": "date-time"
               }
            }
         },
...
         "Level": {
            "type": "object",
            "description": "User authorization level (external/admin)",
            "properties": {
               "id": {
                  "type": "integer",
                  "format": "int32"
               },
               "name": {
                  "type": "string",
                  "enum": [
                     "external",
                     "admin"
                  ]
               },
               "created_at": {
                  "type": "string",
                  "format": "date-time"
               },
               "updated_at": {
                  "type": "string",
                  "format": "date-time"
               }
            }
         },
...
      }
   },
...
}